requirements.txt
langchain==1.2.10
langchain-openai==1.1.11
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
